Stop using Aws::AssumeRoleWebIdentityCredentials
I've never been particularly keen on that class' need to read an access token
off disk; it seems like an unnecessary security risk to leave a token laying
around if it doesn't have to, and it complicates matters for people who don't have
an explicit profile set (because we needed to write that file *somewhere*).

It's only 20 lines of code to do it ourselves, so we'll do that instead.